backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
channels==4.1.0
uvicorn==0.30.6

